Measuring cores with PAPI
BenchCast now includes the option to leave some processor cores without launching any application on them.
Initially, these cores are the first ones. If PAPI is used, performance counter values of those cores are reported.
This is the first step for supporting multithreaded applications, and allows BenchCast to measure the behavior of applications already running in the system, in parallel with the applications launched with it.